--- public/css/register.css
@@ -1,59 +1,0 @@
-/*ヘッダー*/
-.header{
-  height: 80px;
-  width: 100%;
-  background-color: aliceblue;
-}
-.logo{
-  margin-left: 30px;
-  font-size: 25px;
-}
-
-
-/*タイトル*/
-.register-content{
-  margin: 50px;
-}
-
-.register-ttl{
-  text-align: center;
-  font-size: 20px;
-}
-
-/*入力リスト*/
-.register-list{
-  margin-top: 20px;
-}
-
-.register-item{
-  width: 100%;
-  height: 45px;
-  font-size: 12px;
-  padding-left: 20px;
-  box-sizing: border-box;
-  background-color: #f1f1f1;
-  border-radius: 5px;
-  border: solid 2px #b7b7b7;
-}
-
-.register-btn {
-    background-color: #2525e0;
-    color: white;
-    border: none;
-}
-
-.register-login{
-  margin: 20px 0;
-    text-align: center;
-    font-size: 70%;
-}
-
-.register-login-message{
-  color: gray;
-  margin-bottom: 5px;
-}
-
-.register-login-btn{
-  color: #2525e0;
-}
-
