+++ resources/views/layouts/registerView.blade.php
@@ -1,0 +1,19 @@
+@extends('layouts.base')
+
+<style>
+th{
+  
+}
+
+
+
+
+
+
+
+
+
+
+
+
+</style>
