+++ routes/web.php
@@ -1,13 +1,16 @@
 <?php
 
 use Illuminate\Support\Facades\Route;
-use App\Http\Controllers\UserController;
 
-//ユーザー新規登録ページ表示
-Route::get('/register', [UserController::class, 'getregister']);
-//ユーザー新規登録処理
-Route::post('/register', [UserController::class, 'postregister']);
-//ログインページ表示
-Route::get('/login', [UserController::class, 'getlogin']);
-//ログイン処理
-Route::post('/login', [UserController::class, 'postlogin']);
+/*
+|--------------------------------------------------------------------------
+| Web Routes
+|--------------------------------------------------------------------------
+|
+| Here is where you can register web routes for your application. These
+| routes are loaded by the RouteServiceProvider within a group which
+| contains the "web" middleware group. Now create something great!
+|
+*/
+
+Route::get('/', [AuthorController::class, 'index']);
